# contracts/openapi/public-service.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SyncRequestCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "trace_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "title": "SyncRequestCreateRequest",
        "type": "object"
      },
      "SyncRequestCreateResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "status": {
            "enum": [
              "PENDING",
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          },
          "trigger_type": {
            "enum": [
              "manual",
              "scheduler",
              "webhook"
            ],
            "title": "Trigger Type",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "source_id",
          "trigger_type",
          "status",
          "created_at",
          "idempotency_key"
        ],
        "title": "SyncRequestCreateResponse",
        "type": "object"
      },
      "SyncRequestStatusResponse": {
        "properties": {
          "applied": {
            "default": false,
            "title": "Applied",
            "type": "boolean"
          },
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "connector_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Result"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "llm_usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Usage"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "stage": {
            "anyOf": [
              {
                "enum": [
                  "connector_fetch",
                  "llm_queue",
                  "llm_parse",
                  "provider_reduce",
                  "result_ready",
                  "applying",
                  "completed",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "stage_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Updated At"
          },
          "status": {
            "enum": [
              "PENDING",
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          },
          "substage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substage"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "trigger_type": {
            "enum": [
              "manual",
              "scheduler",
              "webhook"
            ],
            "title": "Trigger Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "source_id",
          "trigger_type",
          "status",
          "idempotency_key",
          "trace_id",
          "error_code",
          "error_message",
          "metadata",
          "created_at",
          "updated_at"
        ],
        "title": "SyncRequestStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookEnqueueResponse": {
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "WebhookEnqueueResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CalendarDIFF API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sources/{source_id}/sync-requests": {
      "post": {
        "operationId": "create_sync_request_sources__source_id__sync_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRequestCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Sync Request",
        "tags": [
          "input-control-plane"
        ]
      }
    },
    "/sources/{source_id}/webhooks/{provider}": {
      "post": {
        "operationId": "webhook_ingest_sources__source_id__webhooks__provider__post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Event-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Event-Id"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEnqueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Webhook Ingest",
        "tags": [
          "input-control-plane"
        ]
      }
    },
    "/sync-requests/{request_id}": {
      "get": {
        "operationId": "get_sync_request_sync_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRequestStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sync Request",
        "tags": [
          "input-control-plane"
        ]
      }
    }
  }
}